pyproject.toml
[tool.poetry]
name = "spoty_etl"
version = "0.1.0"
description = ""
authors = ["rishis07 <risheric@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
spotipy = "^2.16.1"
python-decouple = "^3.3"
alembic = "^1.4.3"
psycopg2-binary = "^2.8.6"
pandas = "1.0.3"

[tool.poetry.dev-dependencies]
pdbpp = "^0.10.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
